Added a first implementation of resources. Only for playerBullet so far. TODO: make it so the resources need to be created only once and then can be referred to by a name

+++ Entity.cpp
@@ -36,11 +36,11 @@ void entities::Entity::markDeleted() {
     m_deleted = true;
 }
 
-std::pair<double, double> entities::Entity::getPosition() {
+std::pair<float, float> entities::Entity::getPosition() {
     return m_position;
 }
 
-std::pair<double, double> entities::Entity::getDirection() {
+std::pair<float, float> entities::Entity::getDirection() {
     return m_direction;
 }
